made searchbar its own component
