+++ app/api/playlist_routes.py
@@ -9,13 +9,13 @@ playlist_routes = Blueprint('playlists', __name__)
 # @login_required
 def playlists():
     playlists = Playlist.query.all()
-    print(playlists,'---------------------------------')
+    # print(playlists,'---------------------------------')
     return {'playlists': [playlist.to_dict() for playlist in playlists]}
 
 @playlist_routes.route('/<string:search_value>')
 # @login_required
 def playlist(search_value):
-    print(search_value)
+    # print(search_value)
     playlist_search_results = Playlist.query.filter(Playlist.name.ilike(f'%{search_value}%')).all()
-    print('---------------------------------',playlist_search_results,'---------------------------------')
+    # print('---------------------------------',playlist_search_results,'---------------------------------')
     return {'playlists': [playlist.to_dict() for playlist in playlist_search_results]}
